Restful classes changed to implement interfaces with JAX-RS specific annotations
