+++ server-spring/src/main/java/com/devicehive/json/strategies/JsonPolicyApply.java
@@ -1,11 +1,13 @@
 package com.devicehive.json.strategies;
 
 
+import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 import javax.enterprise.util.AnnotationLiteral;
 
+@Inherited
 @Retention(RetentionPolicy.RUNTIME)
 public @interface JsonPolicyApply {
