+++ server-spring/src/main/java/com/devicehive/json/strategies/JsonPolicyDef.java
@@ -1,10 +1,12 @@
 package com.devicehive.json.strategies;
 
 
+import java.lang.annotation.Inherited;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Retention(RetentionPolicy.RUNTIME)
+@Inherited
 public @interface JsonPolicyDef {
 
     Policy[] value();
