+++ server-spring/src/main/java/com/devicehive/resource/AccessKeyResource.java
@@ -1,48 +1,21 @@
 package com.devicehive.resource;
 
-import com.devicehive.auth.HivePrincipal;
-import com.devicehive.configuration.Constants;
-import com.devicehive.configuration.Messages;
-import com.devicehive.resource.converters.SortOrderQueryParamParser;
-import com.devicehive.resource.util.ResponseFactory;
-import com.devicehive.exceptions.HiveException;
 import com.devicehive.json.strategies.JsonPolicyApply;
 import com.devicehive.model.AccessKey;
-import com.devicehive.model.ErrorResponse;
-import com.devicehive.model.User;
-import com.devicehive.model.enums.UserRole;
 import com.devicehive.model.updates.AccessKeyUpdate;
-import com.devicehive.service.AccessKeyService;
-import com.devicehive.service.UserService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
-import java.util.List;
 
 import static com.devicehive.configuration.Constants.*;
-import static com.devicehive.json.strategies.JsonPolicyDef.Policy.*;
-import static javax.ws.rs.core.Response.Status.*;
+import static com.devicehive.json.strategies.JsonPolicyDef.Policy.ACCESS_KEY_PUBLISHED;
 
 /**
- * REST Controller for access keys: <i>/user/{userId}/accesskey</i> See <a href="http://www.devicehive.com/restful/#Reference/AccessKey">DeviceHive
+ * REST Resource for access keys: <i>/user/{userId}/accesskey</i> See <a href="http://www.devicehive.com/restful/#Reference/AccessKey">DeviceHive
  * RESTful API: AccessKey</a> for details.
  */
-@Service
-@Path("/user/{userId}/accesskey")
-public class AccessKeyResource {
-    private static Logger logger = LoggerFactory.getLogger(AccessKeyResource.class);
-
-    @Autowired
-    private UserService userService;
-
-    @Autowired
-    private AccessKeyService accessKeyService;
+public interface AccessKeyResource {
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/AccessKey/list">DeviceHive RESTful API:
@@ -55,30 +28,15 @@ public class AccessKeyResource {
      */
     @GET
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'MANAGE_ACCESS_KEY')")
-    public Response list(@PathParam(USER_ID) String userId, @QueryParam(LABEL) String label,
-                         @QueryParam(LABEL_PATTERN) String labelPattern, @QueryParam(TYPE) Integer type,
-                         @QueryParam(SORT_FIELD) String sortField, @QueryParam(SORT_ORDER) String sortOrderSt,
-                         @QueryParam(TAKE) Integer take, @QueryParam(SKIP) Integer skip) {
-
-        logger.debug("Access key : list requested for userId : {}", userId);
-
-        Long id = getUser(userId).getId();
-
-        boolean sortOrder = SortOrderQueryParamParser.parse(sortOrderSt);
-
-        if (sortField != null && !ID.equalsIgnoreCase(sortField) && !LABEL.equalsIgnoreCase(sortField)) {
-            return ResponseFactory.response(BAD_REQUEST,
-                    new ErrorResponse(BAD_REQUEST.getStatusCode(),
-                            Messages.INVALID_REQUEST_PARAMETERS));
-        } else if (sortField != null) {
-            sortField = sortField.toLowerCase();
-        }
-        List<AccessKey> keyList = accessKeyService.list(id, label, labelPattern, type, sortField, sortOrder, take, skip);
-
-        logger.debug("Access key : insert proceed successfully for userId : {}", userId);
-
-        return ResponseFactory.response(OK, keyList, ACCESS_KEY_LISTED);
-    }
+    Response list(
+            @PathParam(USER_ID) String userId,
+            @QueryParam(LABEL) String label,
+            @QueryParam(LABEL_PATTERN) String labelPattern,
+            @QueryParam(TYPE) Integer type,
+            @QueryParam(SORT_FIELD) String sortField,
+            @QueryParam(SORT_ORDER) String sortOrderSt,
+            @QueryParam(TAKE) Integer take,
+            @QueryParam(SKIP) Integer skip);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/AccessKey/get">DeviceHive RESTful API:
@@ -92,25 +50,9 @@ public class AccessKeyResource {
     @GET
     @Path("/{id}")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'MANAGE_ACCESS_KEY')")
-    public Response get(@PathParam(USER_ID) String userId, @PathParam(ID) long accessKeyId) {
-
-        logger.debug("Access key : get requested for userId : {} and accessKeyId", userId, accessKeyId);
-
-        Long id = getUser(userId).getId();
-        AccessKey result = accessKeyService.get(id, accessKeyId);
-        if (result == null) {
-            logger.debug("Access key : list failed for userId : {} and accessKeyId : {}. Reason: No access key found" +
-                         ".", userId, accessKeyId);
-            return ResponseFactory
-                .response(NOT_FOUND,
-                          new ErrorResponse(NOT_FOUND.getStatusCode(), "Access key not found."));
-        }
-
-        logger.debug("Access key : insert proceed successfully for userId : {} and accessKeyId : {}", userId,
-                     accessKeyId);
-
-        return ResponseFactory.response(OK, result, ACCESS_KEY_LISTED);
-    }
+    Response get(
+            @PathParam(USER_ID) String userId,
+            @PathParam(ID) long accessKeyId);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/AccessKey/insert">DeviceHive RESTful API:
@@ -122,15 +64,9 @@ public class AccessKeyResource {
      */
     @POST
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'MANAGE_ACCESS_KEY')")
-    public Response insert(@PathParam(USER_ID) String userId,
-                           @JsonPolicyApply(ACCESS_KEY_PUBLISHED) AccessKey key) {
-
-        logger.debug("Access key : insert requested for userId : {}", userId);
-        User user = getUser(userId);
-        AccessKey generatedKey = accessKeyService.create(user, key);
-        logger.debug("Access key : insert proceed successfully for userId : {}", userId);
-        return ResponseFactory.response(CREATED, generatedKey, ACCESS_KEY_SUBMITTED);
-    }
+    Response insert(
+            @PathParam(USER_ID) String userId,
+            @JsonPolicyApply(ACCESS_KEY_PUBLISHED) AccessKey key);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/AccessKey/update">DeviceHive RESTful API:
@@ -139,28 +75,15 @@ public class AccessKeyResource {
      * @param userId      User identifier. Use the 'current' keyword to update access key of the current user.
      * @param accessKeyId Access key identifier.
      * @return If successful, this method returns an <a href="http://www.devicehive .com/restful#Reference/AccessKey/">AccessKey</a>
-     *         resource in the response body according to the specification.
+     * resource in the response body according to the specification.
      */
     @PUT
     @Path("/{id}")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'MANAGE_ACCESS_KEY')")
-    public Response update(@PathParam(USER_ID) String userId, @PathParam(ID) Long accessKeyId,
-                           @JsonPolicyApply(ACCESS_KEY_PUBLISHED) AccessKeyUpdate accessKeyUpdate) {
-        logger.debug("Access key : update requested for userId : {}, access key id : {}, access key : {} ", userId,
-                     accessKeyId, accessKeyUpdate);
-
-        Long id = getUser(userId).getId();
-        if (!accessKeyService.update(id, accessKeyId, accessKeyUpdate)) {
-            logger.debug("Access key : update failed for userId : {} and accessKeyId : {}. Reason: No access key " +
-                         "found.", userId, accessKeyId);
-            return ResponseFactory
-                .response(NOT_FOUND, new ErrorResponse(NOT_FOUND.getStatusCode(), Messages.ACCESS_KEY_NOT_FOUND));
-        }
-
-        logger.debug("Access key : update proceed successfully for userId : {}, access key id : {}, access key : {} ",
-                     userId, accessKeyId, accessKeyUpdate);
-        return ResponseFactory.response(NO_CONTENT);
-    }
+    Response update(
+            @PathParam(USER_ID) String userId,
+            @PathParam(ID) Long accessKeyId,
+            @JsonPolicyApply(ACCESS_KEY_PUBLISHED) AccessKeyUpdate accessKeyUpdate);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/AccessKey/delete">DeviceHive RESTful API:
@@ -169,52 +92,12 @@ public class AccessKeyResource {
      * @param userId      User identifier. Use the 'current' keyword to update access key of the current user.
      * @param accessKeyId Access key identifier.
      * @return If successful, this method returns an <a href="http://www.devicehive .com/restful#Reference/AccessKey/">AccessKey</a>
-     *         resource in the response body according to the specification.
+     * resource in the response body according to the specification.
      */
     @DELETE
     @Path("/{id}")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'MANAGE_ACCESS_KEY')")
-    public Response delete(@PathParam(USER_ID) String userId, @PathParam(ID) Long accessKeyId) {
-        logger.debug("Access key : delete requested for userId : {}", userId);
-
-        Long id = getUser(userId).getId();
-        accessKeyService.delete(id, accessKeyId);
-
-        logger.debug("Access key : delete proceed successfully for userId : {} and access key id : {}", userId,
-                     accessKeyId);
-        return ResponseFactory.response(NO_CONTENT);
-
-    }
-
-    private User getUser(String userId) {
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        User currentUser = principal.getUser() != null ? principal.getUser() : principal.getKey().getUser();
-
-        Long id;
-        if (userId.equalsIgnoreCase(Constants.CURRENT_USER)) {
-            return currentUser;
-        } else {
-            try {
-                id = Long.parseLong(userId);
-            } catch (NumberFormatException e) {
-                throw new HiveException(String.format(Messages.BAD_USER_IDENTIFIER, userId), e,
-                                        BAD_REQUEST.getStatusCode());
-            }
-        }
-
-        User result;
-        if (!currentUser.getId().equals(id) && currentUser.isAdmin()) {
-            result = userService.findById(id);
-            if (result == null) {
-                throw new HiveException(Messages.UNAUTHORIZED_REASON_PHRASE, UNAUTHORIZED.getStatusCode());
-            }
-            return result;
-
-        }
-        if (!currentUser.getId().equals(id) && currentUser.getRole().equals(UserRole.CLIENT)) {
-            throw new HiveException(Messages.UNAUTHORIZED_REASON_PHRASE, UNAUTHORIZED.getStatusCode());
-        }
-        result = currentUser;
-        return result;
-    }
+    Response delete(
+            @PathParam(USER_ID) String userId,
+            @PathParam(ID) Long accessKeyId);
 }
