+++ server-spring/src/main/java/com/devicehive/resource/DeviceResource.java
@@ -1,53 +1,21 @@
 package com.devicehive.resource;
 
-import com.devicehive.auth.HivePrincipal;
-import com.devicehive.auth.HiveRoles;
-import com.devicehive.configuration.Messages;
-import com.devicehive.resource.converters.SortOrderQueryParamParser;
-import com.devicehive.resource.util.ResponseFactory;
-import com.devicehive.json.GsonFactory;
-import com.devicehive.json.strategies.JsonPolicyDef;
-import com.devicehive.model.*;
-import com.devicehive.model.updates.DeviceUpdate;
-import com.devicehive.service.DeviceEquipmentService;
-import com.devicehive.service.DeviceService;
-import com.google.common.collect.ImmutableSet;
-import com.google.gson.Gson;
 import com.google.gson.JsonObject;
-import com.google.gson.reflect.TypeToken;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 import static com.devicehive.configuration.Constants.*;
-import static com.devicehive.json.strategies.JsonPolicyDef.Policy.*;
-import static javax.ws.rs.core.Response.Status.*;
 
 /**
  * REST controller for devices: <i>/device</i>. See <a href="http://www.devicehive.com/restful#Reference/Device">DeviceHive
  * RESTful API: Device</a> for details.
  */
-@Service
-@Path("/device")
-public class DeviceResource {
-    private static final Logger logger = LoggerFactory.getLogger(DeviceResource.class);
-
-    @Autowired
-    private DeviceEquipmentService deviceEquipmentService;
-    @Autowired
-    private DeviceService deviceService;
+public interface DeviceResource {
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/Device/list"> DeviceHive RESTful API:
@@ -69,43 +37,19 @@ public class DeviceResource {
      */
     @GET
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'GET_DEVICE')")
-    public Response list(@QueryParam(NAME) String name,
-                         @QueryParam(NAME_PATTERN) String namePattern,
-                         @QueryParam(STATUS) String status,
-                         @QueryParam(NETWORK_ID) Long networkId,
-                         @QueryParam(NETWORK_NAME) String networkName,
-                         @QueryParam(DEVICE_CLASS_ID) Long deviceClassId,
-                         @QueryParam(DEVICE_CLASS_NAME) String deviceClassName,
-                         @QueryParam(DEVICE_CLASS_VERSION) String deviceClassVersion,
-                         @QueryParam(SORT_FIELD) String sortField,
-                         @QueryParam(SORT_ORDER) String sortOrderSt,
-                         @QueryParam(TAKE) @Min(0) @Max(Integer.MAX_VALUE) Integer take,
-                         @QueryParam(SKIP) @Min(0) @Max(Integer.MAX_VALUE) Integer skip) {
-
-        logger.debug("Device list requested");
-
-        boolean sortOrder = SortOrderQueryParamParser.parse(sortOrderSt);
-        if (sortField != null
-            && !NAME.equalsIgnoreCase(sortField)
-            && !STATUS.equalsIgnoreCase(sortField)
-            && !NETWORK.equalsIgnoreCase(sortField)
-            && !DEVICE_CLASS.equalsIgnoreCase(sortField)) {
-            return ResponseFactory.response(Response.Status.BAD_REQUEST,
-                    new ErrorResponse(BAD_REQUEST.getStatusCode(),
-                            Messages.INVALID_REQUEST_PARAMETERS));
-        } else if (sortField != null) {
-            sortField = sortField.toLowerCase();
-        }
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        List<Device> result = deviceService.getList(name, namePattern, status, networkId, networkName, deviceClassId,
-                                                    deviceClassName, deviceClassVersion, sortField, sortOrder, take,
-                                                    skip, principal);
-
-        logger.debug("Device list proceed result. Result list contains {} elems", result.size());
-
-        return ResponseFactory.response(Response.Status.OK, ImmutableSet.copyOf(result), JsonPolicyDef.Policy.DEVICE_PUBLISHED);
-    }
+    Response list(
+            @QueryParam(NAME) String name,
+            @QueryParam(NAME_PATTERN) String namePattern,
+            @QueryParam(STATUS) String status,
+            @QueryParam(NETWORK_ID) Long networkId,
+            @QueryParam(NETWORK_NAME) String networkName,
+            @QueryParam(DEVICE_CLASS_ID) Long deviceClassId,
+            @QueryParam(DEVICE_CLASS_NAME) String deviceClassName,
+            @QueryParam(DEVICE_CLASS_VERSION) String deviceClassVersion,
+            @QueryParam(SORT_FIELD) String sortField,
+            @QueryParam(SORT_ORDER) String sortOrderSt,
+            @QueryParam(TAKE) @Min(0) @Max(Integer.MAX_VALUE) Integer take,
+            @QueryParam(SKIP) @Min(0) @Max(Integer.MAX_VALUE) Integer skip);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/Device/register">DeviceHive RESTful API:
@@ -121,27 +65,9 @@ public class DeviceResource {
     @Path("/{id}")
     @Consumes(MediaType.APPLICATION_JSON)
     @PreAuthorize("hasPermission(null, 'REGISTER_DEVICE')")
-    public Response register(JsonObject jsonObject, @PathParam(ID) String deviceGuid) {
-        logger.debug("Device register method requested. Guid : {}", deviceGuid);
-
-        Gson mainGson = GsonFactory.createGson(DEVICE_SUBMITTED);
-        DeviceUpdate device = mainGson.fromJson(jsonObject, DeviceUpdate.class);
-        device.setGuid(new NullableWrapper<>(deviceGuid));
-        Gson gsonForEquipment = GsonFactory.createGson();
-        Set<Equipment> equipmentSet = gsonForEquipment.fromJson(
-            jsonObject.get(EQUIPMENT),
-            new TypeToken<HashSet<Equipment>>() {
-            }.getType());
-
-        if (equipmentSet != null) {
-            equipmentSet.remove(null);
-        }
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        deviceService.deviceSaveAndNotify(device, equipmentSet, principal);
-        logger.debug("Device register finished successfully. Guid : {}", deviceGuid);
-
-        return ResponseFactory.response(Response.Status.NO_CONTENT);
-    }
+    Response register(
+            JsonObject jsonObject,
+            @PathParam(ID) String deviceGuid);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/Device/get">DeviceHive RESTful API:
@@ -154,20 +80,8 @@ public class DeviceResource {
     @GET
     @Path("/{id}")
     @PreAuthorize("hasAnyRole('CLIENT', 'DEVICE', 'ADMIN', 'KEY') and hasPermission(null, 'GET_DEVICE')")
-    public Response get(@PathParam(ID) String guid) {
-        logger.debug("Device get requested. Guid {}", guid);
-
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        Device device = deviceService.getDeviceWithNetworkAndDeviceClass(guid, principal);
-
-        if (principal.getRole().equals(HiveRoles.DEVICE)) {
-            logger.debug("Device get proceed successfully. Guid {}", guid);
-            return ResponseFactory.response(Response.Status.OK, device, DEVICE_PUBLISHED_DEVICE_AUTH);
-        }
-        logger.debug("Device get proceed successfully. Guid {}", guid);
-        return ResponseFactory.response(Response.Status.OK, device, DEVICE_PUBLISHED);
-    }
+    Response get(
+            @PathParam(ID) String guid);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/Device/delete">DeviceHive RESTful API:
@@ -179,24 +93,8 @@ public class DeviceResource {
     @DELETE
     @Path("/{id}")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY', 'DEVICE') and hasPermission(null, 'REGISTER_DEVICE')")
-    public Response delete(@PathParam(ID) String guid) {
-
-        logger.debug("Device delete requested");
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        final Device device = deviceService.findByGuidWithPermissionsCheck(guid, principal);
-        if (device == null || !guid.equals(device.getGuid())) {
-            logger.debug("No device found for guid : {}", guid);
-            return ResponseFactory
-                    .response(NOT_FOUND, new ErrorResponse(NOT_FOUND.getStatusCode(),
-                                    String.format(Messages.DEVICE_NOT_FOUND, guid)));
-        }
-
-        deviceService.deleteDevice(guid, principal);
-
-        logger.debug("Device with id = {} is deleted", guid);
-
-        return ResponseFactory.response(NO_CONTENT);
-    }
+    Response delete(
+            @PathParam(ID) String guid);
 
     /**
      * Implementation of <a href="http://www.devicehive.com/restful#Reference/Device/equipment">DeviceHive RESTful API:
@@ -214,17 +112,8 @@ public class DeviceResource {
     @GET
     @Path("/{id}/equipment")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'GET_DEVICE_STATE')")
-    public Response equipment(@PathParam(ID) String guid) {
-        logger.debug("Device equipment requested for device {}", guid);
-
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        Device device = deviceService.getDeviceWithNetworkAndDeviceClass(guid, principal);
-        List<DeviceEquipment> equipments = deviceEquipmentService.findByFK(device);
-
-        logger.debug("Device equipment request proceed successfully for device {}", guid);
-
-        return ResponseFactory.response(OK, equipments, DEVICE_EQUIPMENT_SUBMITTED);
-    }
+    Response equipment(
+            @PathParam(ID) String guid);
 
     /**
      * Gets current state of device equipment. The equipment state is tracked by framework and it could be updated by
@@ -238,25 +127,7 @@ public class DeviceResource {
     @GET
     @Path("/{id}/equipment/{code}")
     @PreAuthorize("hasAnyRole('CLIENT', 'ADMIN', 'KEY') and hasPermission(null, 'GET_DEVICE_STATE')")
-    public Response equipmentByCode(@PathParam(ID) String guid,
-                                    @PathParam(CODE) String code) {
-
-        logger.debug("Device equipment by code requested");
-        HivePrincipal principal = (HivePrincipal) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        Device device = deviceService.getDeviceWithNetworkAndDeviceClass(guid, principal);
-
-        DeviceEquipment equipment = deviceEquipmentService.findByCodeAndDevice(code, device);
-        if (equipment == null) {
-            logger.debug("No device equipment found for code : {} and guid : {}", code, guid);
-            return ResponseFactory
-                .response(NOT_FOUND,
-                          new ErrorResponse(NOT_FOUND.getStatusCode(),
-                                            String.format(Messages.DEVICE_NOT_FOUND, guid)));
-        }
-        logger.debug("Device equipment by code proceed successfully");
-
-        return ResponseFactory.response(OK, equipment, DEVICE_EQUIPMENT_SUBMITTED);
-    }
-
-
+    Response equipmentByCode(
+            @PathParam(ID) String guid,
+            @PathParam(CODE) String code);
 }
