+++ server-spring/src/main/java/com/devicehive/resource/EquipmentResource.java
@@ -1,20 +1,10 @@
 package com.devicehive.resource;
 
-
-import com.devicehive.configuration.Messages;
-import com.devicehive.resource.util.ResponseFactory;
 import com.devicehive.json.strategies.JsonPolicyApply;
 import com.devicehive.json.strategies.JsonPolicyDef;
 import com.devicehive.model.Equipment;
-import com.devicehive.model.ErrorResponse;
 import com.devicehive.model.updates.EquipmentUpdate;
-import com.devicehive.service.DeviceClassService;
-import com.devicehive.service.EquipmentService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.stereotype.Service;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -22,21 +12,8 @@ import javax.ws.rs.core.Response;
 
 import static com.devicehive.configuration.Constants.DEVICE_CLASS_ID;
 import static com.devicehive.configuration.Constants.ID;
-import static com.devicehive.json.strategies.JsonPolicyDef.Policy.EQUIPMENTCLASS_SUBMITTED;
-import static com.devicehive.json.strategies.JsonPolicyDef.Policy.EQUIPMENT_PUBLISHED;
-import static javax.ws.rs.core.Response.Status.*;
 
-@Service
-@Path("/device/class/{deviceClassId}/equipment")
-public class EquipmentResource {
-    private static final Logger logger = LoggerFactory.getLogger(EquipmentResource.class);
-
-    @Autowired
-    private DeviceClassService deviceClassService;
-
-    @Autowired
-    private EquipmentService equipmentService;
-
+public interface EquipmentResource {
 
     /**
      * Gets current state of device equipment. <code> [ { "id":1, "timestamp": "1970-01-01 00:00:00.0", "parameters":{/
@@ -49,22 +26,9 @@ public class EquipmentResource {
     @GET
     @Path("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public Response getEquipment(@PathParam(DEVICE_CLASS_ID) long classId, @PathParam(ID) long eqId) {
-
-        logger.debug("Device class's equipment get requested");
-        Equipment result = equipmentService.getByDeviceClass(classId, eqId);
-
-        if (result == null) {
-            logger.debug("No equipment with id = {} for device class with id = {} found", eqId, classId);
-            return ResponseFactory.response(NOT_FOUND,
-                                            new ErrorResponse(NOT_FOUND.getStatusCode(),
-                                                              String.format(Messages.EQUIPMENT_NOT_FOUND, eqId,
-                                                                            classId)));
-        }
-        logger.debug("Device class's equipment get proceed successfully");
-
-        return ResponseFactory.response(OK, result, EQUIPMENT_PUBLISHED);
-    }
+    Response getEquipment(
+            @PathParam(DEVICE_CLASS_ID) long classId,
+            @PathParam(ID) long eqId);
 
     /**
      * Adds new equipment type to device class
@@ -74,14 +38,9 @@ public class EquipmentResource {
     @POST
     @PreAuthorize("hasRole('ADMIN')")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response insertEquipment(@PathParam(DEVICE_CLASS_ID) long classId, Equipment equipment) {
-
-        logger.debug("Insert device class's equipment requested");
-        Equipment result = deviceClassService.createEquipment(classId, equipment);
-        logger.debug("New device class's equipment created");
-
-        return ResponseFactory.response(CREATED, result, EQUIPMENTCLASS_SUBMITTED);
-    }
+    Response insertEquipment(
+            @PathParam(DEVICE_CLASS_ID) long classId,
+            Equipment equipment);
 
     /**
      * Updates device class' equipment. None of following parameters are mandatory. Parameters, if left unspecified,
@@ -101,26 +60,10 @@ public class EquipmentResource {
     @Path("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response updateEquipment(
-        @PathParam(DEVICE_CLASS_ID) long classId,
-        @PathParam(ID) long eqId,
-        @JsonPolicyApply(JsonPolicyDef.Policy.EQUIPMENT_PUBLISHED) EquipmentUpdate equipmentUpdate) {
-
-        logger.debug("Update device class's equipment requested");
-
-        if (!equipmentService.update(equipmentUpdate, eqId, classId)) {
-            logger.debug("Unable to update equipment. Equipment with id = {} for device class with id = {} not found",
-                         eqId, classId);
-            return ResponseFactory.response(NOT_FOUND,
-                                            new ErrorResponse(NOT_FOUND.getStatusCode(),
-                                                              String.format(Messages.EQUIPMENT_NOT_FOUND, eqId,
-                                                                            classId)));
-        }
-
-        logger.debug("Update device class's equipment finished successfully");
-
-        return ResponseFactory.response(NO_CONTENT);
-    }
+    Response updateEquipment(
+            @PathParam(DEVICE_CLASS_ID) long classId,
+            @PathParam(ID) long eqId,
+            @JsonPolicyApply(JsonPolicyDef.Policy.EQUIPMENT_PUBLISHED) EquipmentUpdate equipmentUpdate);
 
     /**
      * Will cascade deletes specified equipment and all data for this equipment for all devise of this type.
@@ -133,14 +76,9 @@ public class EquipmentResource {
     @Path("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
     @Consumes(MediaType.APPLICATION_JSON)
-    public Response deleteEquipment(@PathParam(DEVICE_CLASS_ID) long classId, @PathParam(ID) long eqId) {
-
-        logger.debug("Delete device class's equipment requested");
-        equipmentService.delete(eqId, classId);
-        logger.debug("Delete device class's equipment finished");
-
-        return ResponseFactory.response(NO_CONTENT);
-    }
+    Response deleteEquipment(
+            @PathParam(DEVICE_CLASS_ID) long classId,
+            @PathParam(ID) long eqId);
 
     /**
      * Gets current state of device equipment. <code> [ { "id":1, "timestamp": "1970-01-01 00:00:00.0", "parameters":{/
@@ -149,9 +87,5 @@ public class EquipmentResource {
      */
     @GET
     @PreAuthorize("hasRole('ADMIN')")
-    public Response getEquipment() {
-        return ResponseFactory.response(METHOD_NOT_ALLOWED, new ErrorResponse(METHOD_NOT_ALLOWED.getStatusCode(),
-                                                                              METHOD_NOT_ALLOWED.getReasonPhrase()));
-    }
-
+    Response getEquipment();
 }
